fix: handled merged conflicts
