+++ core/src/main/java/dk/alexandra/fresco/outsourcing/client/jno/ReconstructClientInput.java
@@ -8,10 +8,13 @@ import dk.alexandra.fresco.framework.builder.numeric.field.FieldElement;
 import dk.alexandra.fresco.framework.util.Pair;
 import dk.alexandra.fresco.framework.value.SInt;
 import dk.alexandra.fresco.lib.common.math.AdvancedNumeric;
-import dk.alexandra.fresco.outsourcing.client.AbstractClientBase;
-
+import dk.alexandra.fresco.outsourcing.utils.GenericUtils;
 import java.math.BigInteger;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.SortedMap;
 import java.util.stream.Collectors;
 
 public class ReconstructClientInput implements Computation<Map<Integer, List<SInt>>, ProtocolBuilderNumeric> {
@@ -104,7 +107,8 @@ public class ReconstructClientInput implements Computation<Map<Integer, List<SIn
             List<DRes<SInt>> tShares = sharedPayloads.stream().map(cur -> cur.getT()).collect(Collectors.toList());
             List<DRes<SInt>> kShares = sharedPayloads.stream().map(cur -> cur.getK()).collect(Collectors.toList());
             List<DRes<SInt>> rShares = sharedPayloads.stream().map(cur -> cur.getR()).collect(Collectors.toList());
-            List<List<DRes<SInt>>> xShares = AbstractClientBase.transpose(sharedPayloads.stream().map(cur -> cur.getX()).collect(Collectors.toList()));
+            List<List<DRes<SInt>>> xShares =
+                GenericUtils.transpose(sharedPayloads.stream().map(cur -> cur.getX()).collect(Collectors.toList()));
             DRes<SInt> t = AdvancedNumeric.using(builder).sum(tShares);
             DRes<SInt> k = AdvancedNumeric.using(builder).sum(kShares);
             DRes<SInt> r = AdvancedNumeric.using(builder).sum(rShares);
